Fix CV drag and drop
